+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-23 10:21:33
-// Commit: test(core): implement widget integration tests for core
+// Updated at: 2026-08-23 10:42:31
+// Commit: test(core): add unit tests for core business logic
 
 class coreComponentState {
-  final String commitRef = '3824';
-  final String lastMessage = 'test(core): implement widget integration tests for core';
-  final DateTime timestamp = DateTime.parse('2026-08-23T10:21:33+05:00');
+  final String commitRef = '3828';
+  final String lastMessage = 'test(core): add unit tests for core business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-23T10:42:31+05:00');
 }